Refactored the construction of a new game
Changelog:
Player: Added boolean, current, this will make easy for us if we want to
represent our player in a player view object to detmine if it's his turn
and then maybe higlight him.

Game: Added a new method newGame. Will be used when first constructing a
game, but can also be use for "restart", without needing to make another
object. newGame only resets values that has to be reset, it leaves
worldMap and battle unchanged.

IGame: Added newgame.